pyproject.toml
[tool.poetry]
name = "mentorship-streamlit"
description = ""
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
streamlit = "^1.34.0"
mysqlclient = "^2.2.4"
SQLAlchemy = "^2.0.36"
plotly = "^5.24.0"
pandas = "^2.2.2"
numpy = "^2.1.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
duckdb = "^1.1.3"
streamlit-code-editor = "^0.1.21"
polars = {extras = ["pyarrow"], version = "^1.18.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
